spotify-profile-backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pandas